Don't avoid elements of hidden sets
It fixes one of the problems in sep-pairs. It probably won't matter anymore for
correctness once we revise overlap calculations, but it will give better error
messages.
